requirements.txt
gradio==4.44.1
flask==2.3.3
flask-cors==4.0.0
gradio-client==0.7.1
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
gtts==2.4.0
pydantic==2.10.6
requests==2.31.0
